requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-tavily
pymongo
gradio
python-dotenv
requests
pypdf               
